+++ Source/Game/src/Game/Elements/Ball.cpp
@@ -41,14 +41,14 @@ void cBall::VInitialize(const cGameElementDef & def )
 		Log_Write_L1(ILogger::LT_DEBUG, cString(100, "Random Generator created for Ball with seed %u", m_pRandomGenerator->GetRandomSeed()));
 	}
 	
-	m_vSpeed = cVector3(m_pRandomGenerator->Random(5,10), m_pRandomGenerator->Random(5,10), 0.0f);
+	m_vSpeed = cVector3(static_cast<float>(m_pRandomGenerator->Random(5,10)), static_cast<float>(m_pRandomGenerator->Random(5,10)), 0.0f);
 }
 
 // *****************************************************************************
 void cBall::OnRestart()
 {
 	cPongGameElement::OnRestart();
-	m_vSpeed = cVector3(m_pRandomGenerator->Random(5,10), m_pRandomGenerator->Random(5,10), 0.0f);
+	m_vSpeed = cVector3(static_cast<float>(m_pRandomGenerator->Random(5,10)), static_cast<float>(m_pRandomGenerator->Random(5,10)), 0.0f);
 	
 	int iSpeedDirection = m_pRandomGenerator->Random(2);
 	if (iSpeedDirection == 1)
